Add blog post about git's autocrlf
I always wanted to write a "considered harmful" article, and when is the
best time if not Halloween itself?
Such articles are clickbait and usually have a bad take.
Maybe I just added myself now to this illustrious circle of bad takers.
But maybe it's even useful.
So best way to figure that out:
Read the article and judge for yourself!